# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leave Planning Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  admin: "admin",
  user: "user",
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  role: UserRole;
  mustChangePassword: boolean;
  createdAt: string;
}

export interface LoginResponse {
  user: User;
  message: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface UpdateProfileBody {
  name: string;
}

export type CreateUserBodyRole =
  (typeof CreateUserBodyRole)[keyof typeof CreateUserBodyRole];

export const CreateUserBodyRole = {
  admin: "admin",
  user: "user",
} as const;

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
}

export interface CreateLeaveBody {
  startDate: string;
  endDate: string;
  reason: string;
}

export interface UpdateLeaveBody {
  startDate?: string;
  endDate?: string;
  reason?: string;
}

export interface LeaveWithUser {
  id: string;
  userId: string;
  startDate: string;
  endDate: string;
  reason: string;
  createdAt: string;
  userName: string;
  userEmail: string;
}

export interface DashboardStats {
  totalMembers: number;
  leavesToday: number;
  upcomingLeaves: number;
  monthlyLeaves: number;
}