# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leave Planning Tracker API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  user: zod.object({
    id: zod.string(),
    name: zod.string(),
    email: zod.string(),
    role: zod.enum(["admin", "user"]),
    mustChangePassword: zod.boolean(),
    createdAt: zod.string(),
  }),
  message: zod.string(),
});

/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["admin", "user"]),
  mustChangePassword: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Change password
 */
export const ChangePasswordBody = zod.object({
  currentPassword: zod.string(),
  newPassword: zod.string(),
});

export const ChangePasswordResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Update profile name
 */
export const UpdateProfileBody = zod.object({
  name: zod.string(),
});

export const UpdateProfileResponse = zod.object({
  id: zod.string(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["admin", "user"]),
  mustChangePassword: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary List all team members (admin only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.string(),
  name: zod.string(),
  email: zod.string(),
  role: zod.enum(["admin", "user"]),
  mustChangePassword: zod.boolean(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Add a team member (admin only)
 */
export const CreateUserBody = zod.object({
  name: zod.string(),
  email: zod.string(),
  password: zod.string(),
  role: zod.enum(["admin", "user"]),
});

/**
 * @summary Remove a team member (admin only)
 */
export const DeleteUserParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary List leaves (all for admin, own for user)
 */
export const ListLeavesResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  startDate: zod.string(),
  endDate: zod.string(),
  reason: zod.string(),
  createdAt: zod.string(),
  userName: zod.string(),
  userEmail: zod.string(),
});
export const ListLeavesResponse = zod.array(ListLeavesResponseItem);

/**
 * @summary Create a leave
 */
export const CreateLeaveBody = zod.object({
  startDate: zod.string(),
  endDate: zod.string(),
  reason: zod.string(),
});

/**
 * @summary List all team leaves for calendar view
 */
export const ListAllLeavesResponseItem = zod.object({
  id: zod.string(),
  userId: zod.string(),
  startDate: zod.string(),
  endDate: zod.string(),
  reason: zod.string(),
  createdAt: zod.string(),
  userName: zod.string(),
  userEmail: zod.string(),
});
export const ListAllLeavesResponse = zod.array(ListAllLeavesResponseItem);

/**
 * @summary Update a leave
 */
export const UpdateLeaveParams = zod.object({
  id: zod.coerce.string(),
});

export const UpdateLeaveBody = zod.object({
  startDate: zod.string().optional(),
  endDate: zod.string().optional(),
  reason: zod.string().optional(),
});

export const UpdateLeaveResponse = zod.object({
  id: zod.string(),
  userId: zod.string(),
  startDate: zod.string(),
  endDate: zod.string(),
  reason: zod.string(),
  createdAt: zod.string(),
  userName: zod.string(),
  userEmail: zod.string(),
});

/**
 * @summary Delete a leave
 */
export const DeleteLeaveParams = zod.object({
  id: zod.coerce.string(),
});

/**
 * @summary Get dashboard stats
 */
export const GetStatsResponse = zod.object({
  totalMembers: zod.number(),
  leavesToday: zod.number(),
  upcomingLeaves: zod.number(),
  monthlyLeaves: zod.number(),
});