# lib/api-zod/src/generated/types/createLeaveBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leave Planning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateLeaveBody {
  startDate: string;
  endDate: string;
  reason: string;
}