# lib/api-zod/src/generated/types/createUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leave Planning Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateUserBodyRole } from "./createUserBodyRole";

export interface CreateUserBody {
  name: string;
  email: string;
  password: string;
  role: CreateUserBodyRole;
}