# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leave Planning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalMembers: number;
  leavesToday: number;
  upcomingLeaves: number;
  monthlyLeaves: number;
}