# lib/api-zod/src/generated/types/leaveWithUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Leave Planning Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface LeaveWithUser {
  id: string;
  userId: string;
  startDate: string;
  endDate: string;
  reason: string;
  createdAt: string;
  userName: string;
  userEmail: string;
}